Pesanan fix

+++ backend-chatbot/database/migrations/2023_10_21_034634_create_order_detail_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->unsignedBigInteger('id_promo')->nullable();
             $table->foreign('id_promo')->references('id')->on('promos');
             $table->decimal('promo_amount')->nullable();
-            $table->decimal('price')->nullable();
+            $table->double('price')->nullable();
             $table->integer('quantity');
             $table->softDeletes();
             $table->timestamps();
